requirements.txt
kafka-python~=2.0.2
streamlit~=1.35.0
pandas~=2.2.0
python-dotenv~=1.0.1
pytest~=8.2.0
